Use conditional compilation
